+++ RedPandaIDE/translations/RedPandaIDE_pt_BR.ts
@@ -1407,12 +1407,12 @@
         <translation type="vanished">O texto a ser removido excede o limite de caracteres!</translation>
     </message>
     <message>
-        <location line="+2909"/>
+        <location line="+2916"/>
         <source>Print Document</source>
         <translation>Imprimir documento</translation>
     </message>
     <message>
-        <location line="+743"/>
+        <location line="+750"/>
         <location line="+25"/>
         <location line="+46"/>
         <source>Ctrl+click for more info</source>
@@ -1460,13 +1460,13 @@
         <translation>Apenas leitura</translation>
     </message>
     <message>
-        <location line="-5421"/>
+        <location line="-5435"/>
         <location line="+506"/>
         <source>Error Load File</source>
         <translation type="unfinished">Erro ao carregar arquivo</translation>
     </message>
     <message>
-        <location line="+1553"/>
+        <location line="+1560"/>
         <source>hex: %1</source>
         <translation type="unfinished"></translation>
     </message>
@@ -1591,18 +1591,28 @@
     </message>
     <message>
         <location line="+9"/>
-        <location line="+67"/>
+        <location line="+81"/>
         <source>Use background color</source>
         <translation>Usar cor de fundo</translation>
     </message>
     <message>
-        <location line="-60"/>
-        <location line="+67"/>
+        <location line="-74"/>
+        <location line="+81"/>
         <source>Use editor&apos;s color scheme</source>
         <translation>Usar esquema de cores do editor</translation>
     </message>
     <message>
-        <location line="-45"/>
+        <location line="-74"/>
+        <source>Copy with line number</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="+7"/>
+        <source>Recalc line number</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="+22"/>
         <location line="+67"/>
         <source>Color scheme</source>
         <translation>Esquema de cores</translation>
@@ -2359,12 +2369,12 @@
         <translation>Configuração</translation>
     </message>
     <message>
-        <location line="+51"/>
+        <location line="+19"/>
         <source>Font:</source>
         <translation>Fonte:</translation>
     </message>
     <message>
-        <location line="+79"/>
+        <location line="+73"/>
         <source>Customize</source>
         <translation type="unfinished"></translation>
     </message>
@@ -2379,33 +2389,72 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="-13"/>
+        <location line="+221"/>
+        <source>OpenType font variation</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="+13"/>
+        <source>Italic [ital]</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="+7"/>
+        <source>Width [wdth]</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="+7"/>
+        <source>Weight [wght]</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="+7"/>
+        <location line="+49"/>
+        <location line="+7"/>
+        <location line="+7"/>
+        <location line="+7"/>
+        <source>Reset</source>
+        <translation type="unfinished">Reset</translation>
+    </message>
+    <message>
+        <location line="-63"/>
+        <source>Slant [slnt]</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="+7"/>
+        <source>Optical size [opsz]</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="-282"/>
         <location line="+3"/>
         <source>Remove custom theme</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="+113"/>
+        <location line="+144"/>
         <source>*Needs restart</source>
         <translation>*Necessário reiniciar</translation>
     </message>
     <message>
-        <location line="-80"/>
+        <location line="+23"/>
         <source>Theme:</source>
         <translation>Tema:</translation>
     </message>
     <message>
-        <location line="-46"/>
+        <location line="-180"/>
         <source>Create a customized copy</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="+101"/>
+        <location line="+187"/>
         <source>Language:</source>
         <translation>Idioma:</translation>
     </message>
     <message>
-        <location line="-48"/>
+        <location line="-212"/>
         <source>Font Size:</source>
         <translation>Tamanho da fonte:</translation>
     </message>
@@ -2414,12 +2463,12 @@
         <translation type="vanished">Usar tema personalizado</translation>
     </message>
     <message>
-        <location line="-174"/>
+        <location line="+78"/>
         <source>Icon Set:</source>
         <translation>Conjunto de ícones:</translation>
     </message>
     <message>
-        <location line="+199"/>
+        <location line="+25"/>
         <source>Use custom icon set</source>
         <translation>Usar conjunto de ícones personalizado</translation>
     </message>
@@ -2444,7 +2493,7 @@
         <translation>Chinês tradicional</translation>
     </message>
     <message>
-        <location filename="../settingsdialog/environmentappearancewidget.ui" line="+71"/>
+        <location filename="../settingsdialog/environmentappearancewidget.ui" line="-32"/>
         <source>Icon Zoom:</source>
         <translation type="unfinished"></translation>
     </message>
@@ -4287,7 +4336,7 @@
     <name>MainWindow</name>
     <message>
         <location filename="../mainwindow.ui" line="+14"/>
-        <location filename="../mainwindow.cpp" line="+1330"/>
+        <location filename="../mainwindow.cpp" line="+1325"/>
         <source>Red Panda C++</source>
         <translation>Red Panda C++</translation>
     </message>
@@ -8427,7 +8476,7 @@
 <context>
     <name>QApplication</name>
     <message>
-        <location filename="../main.cpp" line="+467"/>
+        <location filename="../main.cpp" line="+475"/>
         <source>Error</source>
         <translation>Erro</translation>
     </message>
@@ -8451,7 +8500,7 @@
         <location filename="../autolinkmanager.cpp" line="+54"/>
         <location line="+16"/>
         <location line="+21"/>
-        <location filename="../settings.cpp" line="+3990"/>
+        <location filename="../settings.cpp" line="+4016"/>
         <location filename="../widgets/ojproblemsetmodel.cpp" line="-403"/>
         <location line="+63"/>
         <source>Can&apos;t open file &apos;%1&apos; for read.</source>
@@ -8702,7 +8751,7 @@
     </message>
     <message>
         <location line="+6"/>
-        <location filename="../main.cpp" line="-234"/>
+        <location filename="../main.cpp" line="-241"/>
         <location line="+7"/>
         <source>Error</source>
         <translation>Erro</translation>
@@ -8740,7 +8789,7 @@
     </message>
     <message>
         <location filename="../editorlist.cpp" line="+178"/>
-        <location filename="../mainwindow.cpp" line="+3237"/>
+        <location filename="../mainwindow.cpp" line="+2017"/>
         <source>Save</source>
         <translation>Salvar</translation>
     </message>
@@ -8761,7 +8810,7 @@
         <translation>Impossível gravar arquivo de configurações %1</translation>
     </message>
     <message>
-        <location line="+178"/>
+        <location line="+185"/>
         <source>Can&apos;t load autolink settings</source>
         <translation>Impossível carregar configurações para autolink</translation>
     </message>